backend/requirements.txt
fastapi==0.115.6
httpx==0.28.1
pydantic-settings==2.7.1
pytest==8.3.4
PyYAML==6.0.2
uvicorn[standard]==0.34.0
